+++ src/FortniteReplayReader/Models/PlayerData.cs
@@ -105,7 +105,7 @@ namespace FortniteReplayReader.Models
         public FRepMovement? ReplicatedMovement { get; set; }
         public float? ReplicatedWorldTimeSeconds { get; set; }
 
-        public float? ReplicatedWorldTimeSecondsDouble { get; set; }
+        public double? ReplicatedWorldTimeSecondsDouble { get; set; }
         
         public float? LastUpdateTime { get; set; }
